requirements.txt
matplotlib
tkinter
torch
torchvision
opencv-python
numpy
pillow
customtkinter
torchmetrics